printf %d requires an int.

+++ trunk/test/packtest.c
@@ -7,9 +7,9 @@
 
 int main(void)
 {
-    printf("sizeof(XMSG) == %d (should be %d)\n", sizeof(XMSG), 238);
+    printf("sizeof(XMSG) == %d (should be %d)\n", (int) sizeof(XMSG), 238);
     printf("sizeof(struct _stamp) == %d (should be %d)\n",
-      sizeof(struct _stamp), 4);
+      (int) sizeof(struct _stamp), 4);
     return 0;
 }
